New Feature: EVEMon now has a character wallet journal and wallet transactions monitor.

+++ EVEMon.Common/Contract.cs
@@ -178,7 +178,7 @@ namespace EVEMon.Common
         /// <summary>
         /// Gets the volume.
         /// </summary>
-        public double Volume { get; private set; }
+        public decimal Volume { get; private set; }
 
         /// <summary>
         /// Gets the issuer.
